tests

+++ .scripts/javascript/minify.mjs
@@ -1,6 +1,7 @@
 import { basename } from 'node:path';
 import { minify } from 'terser';
 import { readFile, writeFile } from 'node:fs/promises';
+import { existsSync } from 'node:fs';
 import { logger } from '../utils/logger.mjs';
 
 /**
@@ -10,6 +11,9 @@ import { logger } from '../utils/logger.mjs';
  * @returns {Promise<void>}
  */
 async function minifyJs(file) {
+  if (!existsSync(file)) {
+    throw new Error(`File ${file} doesn't exist`);
+  }
   const fileContent = await readFile(file, { encoding: 'utf8' });
   const content = await minify(fileContent, { sourceMap: false, format: { comments: false } });
   await writeFile(file.replace('.js', '.min.js'), content.code, { encoding: 'utf8', mode: 0o644 });
